Move 1846 draft to step mechanism (#946)
* Move 1846 draft to step mechanism

* Minor tidy up

* Use entity properly

* Tidy up player order
